backend/requirements.txt
tensorflow
flask
numpy
flask-cors
pillow
tensorflow_hub